# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AttendanceMarkOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AttendanceType"
          }
        },
        "required": [
          "type",
          "created_at"
        ],
        "title": "AttendanceMarkOut",
        "type": "object"
      },
...
      "AttendanceType": {
        "enum": [
          "ENTRY",
          "EXIT"
        ],
        "title": "AttendanceType",
        "type": "string"
      },
...
      "Body_recognize_image_api_reconocimiento_image_post": {
        "properties": {
          "camera_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.\\-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "image": {
            "contentMediaType": "application/octet-stream",
            "description": "Imagen o frame de cámara (JPEG, PNG, WebP o BMP)",
            "title": "Image",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/RecognitionMode",
            "default": "recognition"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_recognize_image_api_reconocimiento_image_post",
        "type": "object"
      },
...
      "BoundingBoxOut": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBoxOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "EventOut": {
        "properties": {
          "camera_id": {
            "title": "Camera Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "person_name": {
            "title": "Person Name",
            "type": "string"
          },
          "recognized": {
            "title": "Recognized",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "person_id",
          "person_name",
          "recognized",
          "confidence",
          "distance",
          "camera_id",
          "created_at"
        ],
        "title": "EventOut",
        "type": "object"
      },
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "FaceResultOut": {
        "properties": {
          "attendance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttendanceMarkOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "bbox": {
            "$ref": "#/components/schemas/BoundingBoxOut"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "recognized": {
            "title": "Recognized",
            "type": "boolean"
          }
        },
        "required": [
          "recognized",
          "person_id",
          "name",
          "confidence",
          "distance",
          "bbox"
        ],
        "title": "FaceResultOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecognitionMode": {
        "enum": [
          "recognition",
          "attendance"
        ],
        "title": "RecognitionMode",
        "type": "string"
      },
      "RecognitionOut": {
        "properties": {
          "events_recorded": {
            "title": "Events Recorded",
            "type": "integer"
          },
          "faces_detected": {
            "title": "Faces Detected",
            "type": "integer"
          },
          "image_height": {
            "title": "Image Height",
            "type": "integer"
          },
          "image_width": {
            "title": "Image Width",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FaceResultOut"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "faces_detected",
          "image_width",
          "image_height",
          "events_recorded",
          "results"
        ],
        "title": "RecognitionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Cookie HttpOnly que entrega POST /api/auth/login.",
        "in": "cookie",
        "name": "facetrack_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nReconocimiento facial, asistencia y estadísticas con **procesamiento local**.\n\n**Autenticación.** `POST /api/auth/login` entrega una cookie de sesión HttpOnly;\nel navegador la envía sola en las demás llamadas. Todas las rutas la requieren\nsalvo `GET /api/health` y el login. Las marcadas *Solo ADMIN* exigen ese rol.\n\n**Errores.** Siempre `{\"detail\": str, \"code\": str}`; los 5xx agregan\n`request_id` (también en la cabecera `X-Request-ID` de toda respuesta), que\npermite encontrar el detalle en los logs.\n\n**Privacidad.** Las imágenes se procesan en memoria y se descartan; ninguna\nrespuesta incluye embeddings.\n",
    "title": "FaceTrack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/historial": {
      "get": {
        "operationId": "history_api_historial_get",
        "parameters": [
          {
            "in": "query",
            "name": "person_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person Id"
            }
          },
          {
            "description": "true = conocidos, false = desconocidos",
            "in": "query",
            "name": "recognized",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true = conocidos, false = desconocidos",
              "title": "Recognized"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "description": "Un día (hora local)",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Un día (hora local)",
              "title": "Date"
            }
          },
          {
            "description": "Desde (incluido, hora local)",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde (incluido, hora local)",
              "title": "From"
            }
          },
          {
            "description": "Hasta (incluido, hora local)",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta (incluido, hora local)",
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "History",
        "tags": [
          "reconocimiento"
        ]
      }
    },
...
    "/api/reconocimiento/image": {
      "post": {
        "description": "Detecta y reconoce todos los rostros de la imagen, cada uno de forma\nindependiente. Registra eventos (con cooldown) y, en modo `attendance`,\nasistencia. La imagen no se guarda.",
        "operationId": "recognize_image_api_reconocimiento_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_image_api_reconocimiento_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Imagen inválida o corrupta"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Archivo demasiado grande"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Formato no soportado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Recognize Image",
        "tags": [
          "reconocimiento"
        ]
      }
    }
  },
...
}